refactor(chat): 고객센터 챗봇 상태 로직을 훅으로 분리
